refactor(mvc): consolidate jsx seam into jsx-runtime entry
Merge jsx.ts into jsx-runtime.ts (resolved/hostType become module-private)
and drop the index re-exports - the seam is reached explicitly via
'@expressive/mvc/jsx-runtime', keeping the main entry render-agnostic.
